Move UbloxFirmware7Plus class to its own header file.
